# docs/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "API Системы Управления Банковскими Картами"
  description: "REST API для управления банковскими картами с функциями:\n\n- Ауте\
    нтификация и авторизация пользователей\n- Управление картами (создание, просмо\
    тр, блокировка)\n- Переводы между картами\n- Администрирование системы\n\n**Ро\
    ли пользователей:**\n- USER: базовые операции с собственными картами\n- ADMIN:\
    \ полный доступ к системе\n\n**Безопасность:**\n- JWT токены для аутентификаци\
    и\n- Шифрование номеров карт\n- Маскирование чувствительных данных\n"
  contact:
    name: "Abu"
    url: "https://github.com/DesBasito"
    email: "out1of1mind1exception@gmail.com"
  version: "1.0.0"
servers:
- url: "http://localhost:8080/api/v1"
  description: "Локальный сервер разработки"
security:
- Bearer Authentication: []
...
paths:
...
  /card-block-requests:
    post:
      tags:
      - "Запросы на блокировку карт"
      summary: "Создать запрос на блокировку карты"
      description: "Создание запроса на блокировку собственной карты пользователем"
      operationId: "createBlockRequest"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardBlockRequestCreateDto"
        required: true
      responses:
        "403":
          description: "Нет доступа к данной карте"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
        "200":
          description: "Запрос на блокировку создан"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
        "400":
          description: "Некорректные данные или карта уже заблокирована"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
      security:
      - Bearer Authentication: []
  /card-block-requests/{id}/reject:
    post:
      tags:
      - "Запросы на блокировку карт"
      summary: "Отклонить запрос на блокировку (админ)"
      description: "Отклонение запроса на блокировку администратором"
      operationId: "rejectBlockRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "ID запроса на блокировку"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "adminComment"
        in: "query"
        description: "Комментарий администратора (причина отклонения)"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
      security:
      - Bearer Authentication: []
  /card-block-requests/{id}/cancel:
    post:
      tags:
      - "Запросы на блокировку карт"
      summary: "Отменить запрос на блокировку"
      description: "Отмена запроса на блокировку пользователем (только в статусе PENDING)"
      operationId: "cancelBlockRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "ID запроса на блокировку"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
                enum:
                - "100 CONTINUE"
                - "101 SWITCHING_PROTOCOLS"
                - "102 PROCESSING"
                - "103 EARLY_HINTS"
                - "103 CHECKPOINT"
                - "200 OK"
                - "201 CREATED"
                - "202 ACCEPTED"
                - "203 NON_AUTHORITATIVE_INFORMATION"
                - "204 NO_CONTENT"
                - "205 RESET_CONTENT"
                - "206 PARTIAL_CONTENT"
                - "207 MULTI_STATUS"
                - "208 ALREADY_REPORTED"
                - "226 IM_USED"
                - "300 MULTIPLE_CHOICES"
                - "301 MOVED_PERMANENTLY"
                - "302 FOUND"
                - "302 MOVED_TEMPORARILY"
                - "303 SEE_OTHER"
                - "304 NOT_MODIFIED"
                - "305 USE_PROXY"
                - "307 TEMPORARY_REDIRECT"
                - "308 PERMANENT_REDIRECT"
                - "400 BAD_REQUEST"
                - "401 UNAUTHORIZED"
                - "402 PAYMENT_REQUIRED"
                - "403 FORBIDDEN"
                - "404 NOT_FOUND"
                - "405 METHOD_NOT_ALLOWED"
                - "406 NOT_ACCEPTABLE"
                - "407 PROXY_AUTHENTICATION_REQUIRED"
                - "408 REQUEST_TIMEOUT"
                - "409 CONFLICT"
                - "410 GONE"
                - "411 LENGTH_REQUIRED"
                - "412 PRECONDITION_FAILED"
                - "413 PAYLOAD_TOO_LARGE"
                - "413 REQUEST_ENTITY_TOO_LARGE"
                - "414 URI_TOO_LONG"
                - "414 REQUEST_URI_TOO_LONG"
                - "415 UNSUPPORTED_MEDIA_TYPE"
                - "416 REQUESTED_RANGE_NOT_SATISFIABLE"
                - "417 EXPECTATION_FAILED"
                - "418 I_AM_A_TEAPOT"
                - "419 INSUFFICIENT_SPACE_ON_RESOURCE"
                - "420 METHOD_FAILURE"
                - "421 DESTINATION_LOCKED"
                - "422 UNPROCESSABLE_ENTITY"
                - "423 LOCKED"
                - "424 FAILED_DEPENDENCY"
                - "425 TOO_EARLY"
                - "426 UPGRADE_REQUIRED"
                - "428 PRECONDITION_REQUIRED"
                - "429 TOO_MANY_REQUESTS"
                - "431 REQUEST_HEADER_FIELDS_TOO_LARGE"
                - "451 UNAVAILABLE_FOR_LEGAL_REASONS"
                - "500 INTERNAL_SERVER_ERROR"
                - "501 NOT_IMPLEMENTED"
                - "502 BAD_GATEWAY"
                - "503 SERVICE_UNAVAILABLE"
                - "504 GATEWAY_TIMEOUT"
                - "505 HTTP_VERSION_NOT_SUPPORTED"
                - "506 VARIANT_ALSO_NEGOTIATES"
                - "507 INSUFFICIENT_STORAGE"
                - "508 LOOP_DETECTED"
                - "509 BANDWIDTH_LIMIT_EXCEEDED"
                - "510 NOT_EXTENDED"
                - "511 NETWORK_AUTHENTICATION_REQUIRED"
      security:
      - Bearer Authentication: []
  /card-block-requests/{id}/approve:
    post:
      tags:
      - "Запросы на блокировку карт"
      summary: "Одобрить запрос на блокировку (админ)"
      description: "Одобрение запроса на блокировку и блокировка карты"
      operationId: "approveBlockRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "ID запроса на блокировку"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "adminComment"
        in: "query"
        description: "Комментарий администратора"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Запрос одобрен, карта заблокирована"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
        "400":
          description: "Запрос уже обработан"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
      security:
      - Bearer Authentication: []
...
  /cards/{id}:
    get:
      tags:
      - "Карты"
      summary: "Получить карту по ID"
      description: "Получение подробной информации о карте"
      operationId: "getCard"
      parameters:
      - name: "id"
        in: "path"
        description: "ID карты"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "Карта не найдена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
        "200":
          description: "Информация о карте получена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
        "403":
          description: "Нет доступа к карте"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - Bearer Authentication: []
  /cards/my:
    get:
      tags:
      - "Карты"
      summary: "Получить все карты пользователя"
      description: "Получение списка карт текущего пользователя с пагинацией"
      operationId: "getMyCards"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "Список карт получен"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
      security:
      - Bearer Authentication: []
...
  /card-block-requests/{id}:
    get:
      tags:
      - "Запросы на блокировку карт"
      summary: "Получить запрос на блокировку по ID"
      description: "Получение подробной информации о запросе на блокировку"
      operationId: "getBlockRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "ID запроса на блокировку"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
      security:
      - Bearer Authentication: []
  /card-block-requests/my:
    get:
      tags:
      - "Запросы на блокировку карт"
      summary: "Получить мои запросы на блокировку"
      description: "Получение списка запросов на блокировку текущего пользователя"
      operationId: "getMyBlockRequests"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardBlockRequestDto"
      security:
      - Bearer Authentication: []
  /card-block-requests/admin/all:
    get:
      tags:
      - "Запросы на блокировку карт"
      summary: "Получить все запросы на блокировку (админ)"
      description: "Получение всех запросов на блокировку в системе (только для ад\
        министраторов)"
      operationId: "getAllBlockRequests"
      parameters:
      - name: "status"
        in: "query"
        description: "Статус запроса"
        required: false
        schema:
          type: "string"
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardBlockRequestDto"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
...
    CardDto:
      required:
      - "balance"
      - "cardNumber"
      - "createdAt"
      - "expiryDate"
      - "ownerName"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        cardNumber:
          type: "string"
        ownerName:
          type: "string"
        ownerId:
          type: "integer"
          format: "int64"
        expiryDate:
          type: "string"
          format: "date"
        status:
          type: "string"
        type:
          type: "string"
        balance:
          type: "number"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
    CardBlockRequestCreateDto:
      required:
      - "cardId"
      - "reason"
      type: "object"
      properties:
        cardId:
          type: "integer"
          description: "ID карты для блокировки"
          format: "int64"
          example: 1
        reason:
          maxLength: 500
          minLength: 0
          type: "string"
          description: "Причина блокировки"
          example: "Потеря карты"
      description: "Запрос на создание заявки блокировки карты"
    CardBlockRequestDto:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID запроса"
          format: "int64"
        cardId:
          type: "integer"
          description: "ID карты"
          format: "int64"
        cardNumber:
          type: "string"
          description: "Номер карты (маскированный)"
        userId:
          type: "integer"
          description: "ID пользователя"
          format: "int64"
        userName:
          type: "string"
          description: "Имя пользователя"
        reason:
          type: "string"
          description: "Причина блокировки"
        status:
          type: "string"
          description: "Статус запроса"
          enum:
          - "PENDING"
          - "APPROVED"
          - "REJECTED"
        adminComment:
          type: "string"
          description: "Комментарий администратора"
        createdAt:
          type: "string"
          description: "Дата создания запроса"
          format: "date-time"
        processedAt:
          type: "string"
          description: "Дата обработки запроса"
          format: "date-time"
      description: "Запрос на блокировку карты"
...
    Pageable:
      type: "object"
      properties:
        page:
          minimum: 0
          type: "integer"
          format: "int32"
        size:
          minimum: 1
          type: "integer"
          format: "int32"
        sort:
          type: "array"
          items:
            type: "string"
...
    PageableObject:
      type: "object"
      properties:
        offset:
          type: "integer"
          format: "int64"
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: "boolean"
        unpaged:
          type: "boolean"
        pageNumber:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
    SortObject:
      type: "object"
      properties:
        empty:
          type: "boolean"
        unsorted:
          type: "boolean"
        sorted:
          type: "boolean"
...
    Page:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            type: "object"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
    PageCardBlockRequestDto:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/CardBlockRequestDto"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
...
  securitySchemes:
    Bearer Authentication:
      type: "http"
      description: "Enter JWT token in format: Bearer {token}"
      scheme: "bearer"
      bearerFormat: "JWT"